PoPS (Pest or Pathogen Spread) automated management

This function allows for treating multiple species with the same management strategy. Was designed for multiple lineages of the same species that share common hosts and are in the same study area. This functions loop straight through each stochastic realisation with the same treatment.

auto_manage(
  infected_files,
  host_file,
  total_populations_file,
  parameter_means,
  parameter_cov_matrix,
  temp = FALSE,
  temperature_coefficient_file = "",
  precip = FALSE,
  precipitation_coefficient_file = "",
  model_type = "SI",
  latency_period = 0,
  time_step = "month",
  season_month_start = 1,
  season_month_end = 12,
  start_date = "2008-01-01",
  end_date = "2008-12-31",
  use_lethal_temperature = FALSE,
  temperature_file = "",
  lethal_temperature = -12.87,
  lethal_temperature_month = 1,
  mortality_on = FALSE,
  mortality_rate = 0,
  mortality_time_lag = 0,
  management = FALSE,
  treatment_dates = c(""),
  treatments_file = "",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  natural_dir = "NONE",
  anthropogenic_dir = "NONE",
  number_of_iterations = 100,
  number_of_cores = NA,
  pesticide_duration = 0,
  pesticide_efficacy = 1,
  random_seed = NULL,
  output_frequency = "year",
  output_frequency_n = 1,
  movements_file = "",
  use_movements = FALSE,
  cost_per_meter_sq = 1.37,
  budget = 1500000,
  buffer = 600,
  treatment_priority = "equal",
  treatment_rank = c(1),
  selection_method = "Points",
  selection_priority = "group size",
  points = points,
  treatment_efficacy = 1,
  species = c("species1"),
  direction_first = TRUE,
  start_exposed = FALSE,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  deterministic = FALSE,
  establishment_probability = 0.5,
  dispersal_percentage = 0.99,
  quarantine_areas_file = "",
  use_quarantine = FALSE,
  use_spreadrates = FALSE,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1,
  exposed_file = "",
  mask = NULL,
  write_outputs = "None",
  output_folder_path = ""
)

Arguments

infected_files
file path to the infected species files for the start of the simulation

host_file
path to raster files with number of hosts and standard deviation on those estimates can be based in 3 formats (a single file with number of hosts, a single file with 2 layers number of hosts and standard deviation, or two files 1 with number of hosts and the other with standard deviation of those estimates)

total_populations_file
path to raster file with number of total populations of all hosts and non-hosts (must be only hosts if using movement)

parameter_means
A vector of the means of the model parameters (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa)

parameter_cov_matrix
A covariance matrix from the previous years posterior parameter estimation ordered from (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa)

temp
boolean that allows the use of temperature coefficients to modify spread (TRUE or FALSE)

temperature_coefficient_file
path to raster file with temperature coefficient data for the timestep and number of years specified

precip
boolean that allows the use of precipitation coefficients to modify spread (TRUE or FALSE)

precipitation_coefficient_file
path to raster file with precipitation coefficient data for the timestep and number of years specified

model_type
What type of model most represents your sysetm. Options are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI" (Susceptible - Infected/Infested). Default value is "SI".

latency_period
How many times steps does it take to for exposed populations become infected/infested. This is an integer value and must be greater than 0 if model type is SEI.

time_step
how often should spread occur options: ('day', 'week', 'month')

season_month_start
when does spread first start occurring in the year for your pest or pathogen (integer value between 1 and 12)

season_month_end
when does spread end during the year for your pest or pathogen (integer value between 1 and 12)

start_date
date to start the simulation with format ('YYYY_MM_DD')

end_date
date to end the simulation with format ('YYYY_MM_DD')

use_lethal_temperature
a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)

temperature_file
path to raster file with temperature data for minimum temperature

lethal_temperature
the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 and 60)

lethal_temperature_month
the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12

mortality_on
boolean to turn host mortality on and off (TRUE or FALSE)

mortality_rate
rate at which mortality occurs value between 0 and 1

mortality_time_lag
time lag from infection until mortality can occur in years integer >= 1

management
boolean to allow use of management (TRUE or FALSE)

treatment_dates
dates in which to apply treatment list with format ('YYYY_MM_DD') (needs to be the same length as treatment_file and pesticide_duration)

treatments_file
path to raster files with treatment data by dates. Needs to be a list of files the same length as treatment_dates and pesticide_duration.

treatment_method
what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.

natural_kernel_type
what type of dispersal kernel should be used for natural dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic')

anthropogenic_kernel_type
what type of dispersal kernel should be used for anthropogenic dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic')

natural_dir
sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

anthropogenic_dir
sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

number_of_iterations
how many iterations do you want to run to allow the calibration to converge at least 10

number_of_cores
enter how many cores you want to use (default = NA). If not set uses the # of CPU cores - 1. must be an integer >= 1

pesticide_duration
how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment. (needs to be the same length as treatment_dates and treatment_file)

pesticide_efficacy
how effictive is the pesticide at preventing the disease or killing the pest (if this is 0.70 then when applied it successfully treats 70 percent of the plants or animals)

random_seed
sets the random seed for the simulation used for reproducibility

output_frequency
sets when outputs occur either ('year', 'month', 'week', 'day' or 'time step')

output_frequency_n
sets number units ('year', 'month', 'week', 'day' or 'time step') in which to export model results.

movements_file
this is a csv file with columns lon_from, lat_from, lon_to, lat_to, number of animals, and date.

use_movements
this is a boolean to turn on use of the movement module.

cost_per_meter_sq
the cost of treatment per square meter

budget
the total budget to spend on management

buffer
the size of the buffer to include around managed locations

treatment_priority
how to prioritize which of many species to treat options are 'equal' or 'ranked' where equal selects locations based on guidelines from any of the species and rank selects from the species in ranked order and then

treatment_rank
binary 0 or 1 for the species that is the most important

selection_method
the method for determining the management strategy to use. must be one of 'Foci', 'Border', or'Points'.

selection_priority
how to prioritize locations for management must be one of "group size", "host", or "infected"

points
used if selection_method is points

treatment_efficacy
The overall efficacy of the treatment

species
a list of the species names for naming outputs files must be the same length and infected_files

direction_first
boolean to indicate if direction is the first priortity in sorting (if false first sorting priority goes to the selection_method)

start_exposed
Do your initial conditions start as exposed or infected (only used if model_type is "SEI"). Default False. If this is TRUE need to have both an infected_file (this can be a raster of all 0's) and exposed_file

generate_stochasticity
Boolean to indicate whether to use stochasticity in reproductive functions default is TRUE

establishment_stochasticity
Boolean to indicate whether to use stochasticity in establishment functions default is TRUE

movement_stochasticity
Boolean to indicate whether to use stochasticity in movement functions default is TRUE

deterministic
Boolean to indicate whether to use a deterministic dispersal kernel default is FALSE

establishment_probability
Threshold to determine establishment if establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)

dispersal_percentage
Percentage of dispersal used to calculate the bounding box for deterministic dispersal

quarantine_areas_file
path to raster file with quarantine boundaries used in calculating likelihood of quarantine escape if use_quarantine is TRUE

use_quarantine
boolean to indicate whether or not there is a quarantine area if TRUE must pass in a raster file indicating the quarantine areas (default = FALSE)

use_spreadrates
boolean to indicate whether or not to calculate spread rates

use_overpopulation_movements
boolean to indicate whether to use the overpopulation pest movement module (driven by the natural kernel with its scale parameter modified by a coefficient)

overpopulation_percentage
percentage of occupied hosts when the cell is considered to be overpopulated

leaving_percentage
percentage of pests leaving an overpopulated cell

leaving_scale_coefficient
coefficient to multiply scale parameter of the natural kernel (if applicable)

exposed_file
a file with the exposed for the current

mask
Raster file used to provide a mask to remove 0's that are not true negatives from comparisons (e.g. mask out lakes and oceans from statics if modeling terrestrial species). This can also be used to mask out areas that can't be managed in the auto_manage function.

write_outputs
Either c("all_simulations", "summary outputs", or "None"). If not "None" output folder path must be provided.

output_folder_path
this is the full path with either / or \ (e.g., "C:/user_name/desktop/pops_sod_2020_2023/outputs/")

Value

list of infected and susceptible per year