PoPS (Pest or Pathogen Spread) model

A dynamic species distribution model for pest or pathogen spread in forest or agricultural ecosystems. The model is process based meaning that it uses understanding of the effect of weather and other environmental factors on reproduction and survival of the pest/pathogen in order to forecast spread of the pest/pathogen into the future. This function performs a single stochastic realization of the model and is predominantly used for automated tests of model features.

pops(
  infected_file,
  host_file,
  total_populations_file,
  parameter_means,
  parameter_cov_matrix,
  temp = FALSE,
  temperature_coefficient_file = "",
  precip = FALSE,
  precipitation_coefficient_file = "",
  model_type = "SI",
  latency_period = 0,
  time_step = "month",
  season_month_start = 1,
  season_month_end = 12,
  start_date = "2008-01-01",
  end_date = "2008-12-31",
  use_survival_rates = FALSE,
  survival_rate_month = 3,
  survival_rate_day = 15,
  survival_rates_file = "",
  use_lethal_temperature = FALSE,
  temperature_file = "",
  lethal_temperature = -12.87,
  lethal_temperature_month = 1,
  mortality_on = FALSE,
  mortality_rate = 0,
  mortality_time_lag = 0,
  mortality_frequency = "year",
  mortality_frequency_n = 1,
  management = FALSE,
  treatment_dates = c(""),
  treatments_file = "",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  natural_dir = "NONE",
  anthropogenic_dir = "NONE",
  pesticide_duration = c(0),
  pesticide_efficacy = 1,
  random_seed = NULL,
  output_frequency = "year",
  output_frequency_n = 1,
  movements_file = "",
  use_movements = FALSE,
  start_exposed = FALSE,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  dispersal_stochasticity = TRUE,
  establishment_probability = 0.5,
  dispersal_percentage = 0.99,
  quarantine_areas_file = "",
  use_quarantine = FALSE,
  use_spreadrates = FALSE,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1,
  exposed_file = "",
  mask = NULL,
  network_filename = "",
  network_movement = "walk"
)

Arguments

infected_file

Raster file with initial infections. Units for infections are based on data availability and the way the units used for your host file is created (e.g. percent area, # of hosts per cell, etc.).

host_file

path to raster files with number of hosts and standard deviation on those estimates can be based in 3 formats (a single file with number of hosts, a single file with 2 layers number of hosts and standard deviation, or two files 1 with number of hosts and the other with standard deviation of those estimates). The units for this can be of many formats the two most common that we use are either percent area (0 to 100) or # of hosts in the cell. Usually depends on data available and estimation methods.

total_populations_file

path to raster file with number of total populations of all hosts and non-hosts. This depends on how your host data is set up. If host is percent area then this should be a raster with values that are 100 anywhere with host. If host file is # of hosts in a cell then this should be a raster with values that are the max of the host raster any where the # of hosts is greater than 0.

parameter_means

A vector of the means of the model parameters (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, anthropogenic kappa, network_min_distance, and network_max_distance). 1x8 vector.

parameter_cov_matrix

A covariance matrix from the previous years posterior parameter estimation ordered from (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, anthropogenic kappa, network_min_distance, and network_max_distance) Should be 8x8 matrix.

temp

boolean that allows the use of temperature coefficients to modify spread (TRUE or FALSE)

temperature_coefficient_file

path to raster file with temperature coefficient data for the timestep and and time period specified (e.g. if timestep = week and start_date = 2017_01_01 and end_date = 2019_12_31 this file would have 52 * 3 bands = 156 bands with data being weekly precipitation coefficients). We convert raw precipitation values to coefficients that affect the reproduction and survival of the pest all values in the raster are between 0 and 1.

precip

boolean that allows the use of precipitation coefficients to modify spread (TRUE or FALSE)

precipitation_coefficient_file

Raster file with precipitation coefficient data for the timestep and time period specified (e.g. if timestep = week and start_date = 2017_01_01 and end_date = 2019_12_31 this file would have 52 * 3 bands = 156 bands with data being weekly precipitation coefficients). We convert raw precipitation values to coefficients that affect the reproduction and survival of the pest all values in the raster are between 0 and 1.

model_type

What type of model most represents your system. Options are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI" (Susceptible - Infected/Infested). Default value is "SI".

latency_period

How many times steps does it take to for exposed populations become infected/infested. This is an integer value and must be greater than 0 if model type is SEI.

time_step

how often should spread occur options: ('day', 'week', 'month').

season_month_start

when does spread first start occurring in the year for your pest or pathogen (integer value between 1 and 12)

season_month_end

when does spread end during the year for your pest or pathogen (integer value between 1 and 12)

start_date

date to start the simulation with format ('YYYY_MM_DD')

end_date

date to end the simulation with format ('YYYY_MM_DD')

use_survival_rates

boolean to indicate if the model will use survival rates to limit the survival or emergence of overwintering generations.

survival_rate_month

What month do over wintering generations emerge. We suggest using the month before for this parameter as it is when the survival rates raster will be applied.

survival_rate_day

what day should the survival rates be applied

survival_rates_file

Raster file with survival rates from 0 to 1 representing the percentage of emergence for a cell.

use_lethal_temperature

a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)

temperature_file

path to raster file with temperature data for minimum temperature

lethal_temperature

the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 to 60)

lethal_temperature_month

the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12

mortality_on

boolean to turn host mortality on and off (TRUE or FALSE)

mortality_rate

rate at which mortality occurs value between 0 and 1

mortality_time_lag

time lag from infection until mortality can occur in time steps integer >= 1

mortality_frequency

sets the frequency of mortality calculations occur either ('year', 'month', week', 'day', 'time step', or 'every_n_steps')

mortality_frequency_n

sets number of units from mortality_frequency in which to run the mortality calculation if mortality_frequency is 'every_n_steps'. Must be an integer >= 1.

management

boolean to allow use of management (TRUE or FALSE)

treatment_dates

dates in which to apply treatment list with format ('YYYY_MM_DD') (needs to be the same length as treatment_file and pesticide_duration)

treatments_file

path to raster files with treatment data by dates. Needs to be a list of files the same length as treatment_dates and pesticide_duration.

treatment_method

what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.

natural_kernel_type

what type of dispersal kernel should be used for natural dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic')

anthropogenic_kernel_type

what type of dispersal kernel should be used for anthropogenic dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic', 'network')

natural_dir

sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

anthropogenic_dir

sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

pesticide_duration

how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment. (needs to be the same length as treatment_dates and treatment_file)

pesticide_efficacy

how effictive is the pesticide at preventing the disease or killing the pest (if this is 0.70 then when applied it successfully treats 70 percent of the plants or animals)

random_seed

sets the random seed for the simulation used for reproducibility

output_frequency

sets when outputs occur either ('year', 'month', 'week', 'day', 'time step', or 'every_n_steps')

output_frequency_n

sets number of units from output_frequency in which to export model results if mortality_frequency is 'every_n_steps'. Must be an integer >= 1.

movements_file

this is a csv file with columns lon_from, lat_from, lon_to, lat_to, number of animals, and date.

use_movements

this is a boolean to turn on use of the movement module.

start_exposed

Do your initial conditions start as exposed or infected (only used if model_type is "SEI"). Default False. If this is TRUE need to have both an infected_file (this can be a raster of all 0's) and exposed_file

generate_stochasticity

Boolean to indicate whether to use stochasticity in reproductive functions default is TRUE

establishment_stochasticity

Boolean to indicate whether to use stochasticity in establishment functions default is TRUE

movement_stochasticity

Boolean to indicate whether to use stochasticity in movement functions default is TRUE

dispersal_stochasticity

Boolean to indicate whether to use a stochasticity in the dispersal kernel default is TRUE

establishment_probability

Threshold to determine establishment if establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)

dispersal_percentage

Percentage of dispersal used to calculate the bounding box for deterministic dispersal

quarantine_areas_file

path to raster file with quarantine boundaries used in calculating likelihood of quarantine escape if use_quarantine is TRUE

use_quarantine

boolean to indicate whether or not there is a quarantine area if TRUE must pass in a raster file indicating the quarantine areas (default = FALSE)

use_spreadrates

boolean to indicate whether or not to calculate spread rates

use_overpopulation_movements

boolean to indicate whether to use the overpopulation pest movement module (driven by the natural kernel with its scale parameter modified by a coefficient)

overpopulation_percentage

percentage of occupied hosts when the cell is considered to be overpopulated

leaving_percentage

percentage of pests leaving an overpopulated cell

leaving_scale_coefficient

coefficient to multiply scale parameter of the natural kernel (if applicable)

exposed_file

a file with the exposed for the current

mask

Raster file used to provide a mask to remove 0's that are not true negatives from comparisons (e.g. mask out lakes and oceans from statics if modeling terrestrial species). This can also be used to mask out areas that can't be managed in the auto_manage function.

network_filename

entire file path for the network file. Used if anthropogenic_kernel_type = 'network'.

network_movement

What movement type do you want to use in the network kernel either "walk", "jump", or "teleport". "walk" allows dispersing units to leave the network at any cell along the edge. "jump" automatically moves to the nearest node when moving through the network. "teleport" moves from node to node most likely used for airport and seaport networks.

Value

list of infected and susceptible per year