PoPS (Pest or Pathogen Spread) model

Wrapper for pops_model_cpp. A dynamic species distribution model for pest or pathogen spread in forest or agricultural ecosystems. The model is process based meaning that it uses understanding of the effect of weather and other environmental factors on reproduction and survival of the pest/pathogen in order to forecast spread of the pest/pathogen into the future.

pops_model(
  random_seed,
  use_lethal_temperature,
  lethal_temperature,
  lethal_temperature_month,
  infected,
  total_exposed,
  exposed,
  susceptible,
  total_populations,
  total_hosts,
  mortality_on,
  mortality_tracker,
  mortality,
  quarantine_areas,
  treatment_maps,
  treatment_dates,
  pesticide_duration,
  resistant,
  use_movements,
  movements,
  movements_dates,
  weather,
  temperature,
  weather_coefficient,
  res,
  rows_cols,
  time_step,
  reproductive_rate,
  spatial_indices,
  season_month_start_end,
  mortality_rate = 0,
  mortality_time_lag = 2,
  start_date = "2018-01-01",
  end_date = "2018-12-31",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  use_anthropogenic_kernel = FALSE,
  percent_natural_dispersal = 0,
  natural_distance_scale = 21,
  anthropogenic_distance_scale = 0,
  natural_dir = "NONE",
  natural_kappa = 0,
  anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0,
  output_frequency = "year",
  output_frequency_n = 1,
  quarantine_frequency = "year",
  quarantine_frequency_n = 1,
  use_quarantine = FALSE,
  spreadrate_frequency = "year",
  spreadrate_frequency_n = 1,
  mortality_frequency = "year",
  mortality_frequency_n = 1,
  use_spreadrates = FALSE,
  model_type_ = "SI",
  latency_period = 0,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  deterministic = FALSE,
  establishment_probability = 0,
  dispersal_percentage = 0.99,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1
)

Arguments

random_seed
sets the random seed for the simulation used for reproducibility

use_lethal_temperature
a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)

lethal_temperature
the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 and 60)

lethal_temperature_month
the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12

infected
matrix of infected hosts

total_exposed
sum of all exposed cohorts in exposed class for use with "SEI" model type

exposed
vector of matrices of the exposed class for use with "SEI" model type

susceptible
matrix of susceptible hosts

total_populations
matrix of total populations

total_hosts
matrix of all hosts

mortality_on
boolean to turn host mortality on and off (TRUE or FALSE)

mortality_tracker
matrix of 0's to track mortality per year

mortality
matrix to track cumulative mortality

quarantine_areas
areas that are set as quarantined for computing escape from quarantine statistics.

treatment_maps
list of matrices where treatment or management has occurred in a given year

treatment_dates
dates in which to apply treatment list with format ('YYYY_MM_DD') (needs to be the same length as treatment_file and pesticide_duration)

pesticide_duration
how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment. (needs to be the same length as treatment_dates and treatment_file)

resistant
matrix to track resistant population over time

use_movements
this is a boolean to turn on use of the movement module.

movements
a matrix with columns lon_from, lat_from, lon_to, lat_to, number of animals, and date.

movements_dates
this is a list of dates passed as strings in the format 'YYYY-MM-DD'

weather
Boolean that is true if weather is used

temperature
vector of matrices of temperature values used to check against lethal temperature

weather_coefficient
vector of matrices of weather coefficients

res
vector of east/west resolution and north/south resolution

rows_cols
vector of number of rows and columns in the raster files

time_step
how often should spread occur options: ('day', 'week', 'month')

reproductive_rate
number of spores or pest units produced by a single host under optimal weather conditions

spatial_indices
list of all spatial locations with suitable hosts

season_month_start_end
vector of months when spread starts and stops

mortality_rate
rate at which mortality occurs value between 0 and 1

mortality_time_lag
time lag from infection until mortality can occur in time steps integer >= 1

start_date
date to start the simulation with format ('YYYY_MM_DD')

end_date
date to end the simulation with format ('YYYY_MM_DD')

treatment_method
what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.

natural_kernel_type
what type of dispersal kernel should be used for natural dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic')

anthropogenic_kernel_type
what type of dispersal kernel should be used for anthropogenic dispersal. Current dispersal kernel options are ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law', 'hyperbolic secant', 'gamma', 'weibull', 'logistic')

use_anthropogenic_kernel
A boolean that turns on the use of the anthropogenic or long distance dispersal portion of the 2 scale dispersal kernel function

percent_natural_dispersal
what percentage of dispersal is natural range versus anthropogenic range value between 0 and 1

natural_distance_scale
distance scale parameter for natural range dispersal kernel numeric value > 0

anthropogenic_distance_scale
distance scale parameter for anthropogenic range dispersal kernel numeric value > 0

natural_dir
sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

natural_kappa
sets the strength of the natural direction in the von-mises distribution numeric value between 0.01 and 12

anthropogenic_dir
sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')

anthropogenic_kappa
sets the strength of the anthropogenic direction in the von-mises distribution numeric value between 0.01 and 12

output_frequency
sets when outputs occur either ('year', 'month', 'week', 'day', 'time step', or 'every_n_steps')

output_frequency_n
sets number of units from output_frequency in which to export model results if mortality_frequency is 'every_n_steps'. Must be an integer >= 1.

quarantine_frequency
sets how often the quarantine statistics are calculated either ('year', 'month', 'week', 'day' or 'time step') (default is year)

quarantine_frequency_n
sets number units ('year', 'month', 'week', 'day' or 'time step') in which to calculate and export quarantine statistics.

use_quarantine
boolean to indicate whether or not there is a quarantine area if TRUE must pass in a raster file indicating the quarantine areas (default = FALSE)

spreadrate_frequency
sets how often the spread rate statistics are calculated either ('year', 'month', 'week', 'day' or 'time step') (default is year)

spreadrate_frequency_n
sets number units ('year', 'month', 'week', 'day' or 'time step') in which to calculate and export spread rate statistics.

mortality_frequency
sets the frequency of mortality calculations occur either ('year', 'month', week', 'day', 'time step', or 'every_n_steps')

mortality_frequency_n
sets number of units from mortality_frequency in which to run the mortality calculation if mortality_frequency is 'every_n_steps'. Must be an integer >= 1.

use_spreadrates
boolean to indicate whether or not to calculate spread rates

model_type_
What type of model most represents your system. Options are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI" (Susceptible - Infected/Infested). Default value is "SI".

latency_period
How many times steps does it take to for exposed populations become infected/infested. This is an integer value and must be greater than 0 if model type is SEI.

generate_stochasticity
Boolean to indicate whether to use stochasticity in reproductive functions default is TRUE

establishment_stochasticity
Boolean to indicate whether to use stochasticity in establishment functions default is TRUE

movement_stochasticity
Boolean to indicate whether to use stochasticity in movement functions default is TRUE

deterministic
Boolean to indicate whether to use a deterministic dispersal kernel default is FALSE

establishment_probability
Threshold to determine establishment if establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)

dispersal_percentage
Percentage of dispersal used to calculate the bounding box for deterministic dispersal

use_overpopulation_movements
boolean to indicate whether to use the overpopulation pest movement module (driven by the natural kernel with its scale parameter modified by a coefficient)

overpopulation_percentage
percentage of occupied hosts when the cell is considered to be overpopulated

leaving_percentage
percentage of pests leaving an overpopulated cell

leaving_scale_coefficient
coefficient to multiply scale parameter of the natural kernel (if applicable)

Value

list of vector matrices of infected and susceptible hosts per simulated year and associated statistics (e.g. spread rate)